Update europe.json
I noticed that the islands belonging to Spain, France and Italy, were not considered part of those countries as a whole. As a result - for instance - if you would like to highlight a country when a mouseover event occurs over any country, the aforementioned countries are not processed correctly since their islands are not highlighted at the same time that the mainland.

Therefore, I have modified the europe.json file accordingly in order to makes it working in the proper way.